move graph reload logic to the edge component
