+++ community/ui/graph/bubble-graph/bubble-graph.tsx
@@ -1,6 +1,5 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import classNames from 'classnames';
-import { Xwrapper, useXarrow } from 'react-xarrows';
 import { GridGraph, GridGraphProps } from '@teambit/community.ui.graph.grid-graph';
 import { GridNode } from '@teambit/community.entity.graph.grid-graph';
 import { BubbleNode, BubblePayload } from './bubble-node';
@@ -12,22 +11,8 @@ export type BubbleGraphProps = {
 
 export function BubbleGraph({ nodes, Node = BubbleNode, children, nodeClassName, ...rest }: BubbleGraphProps) {
   return (
-    <Xwrapper>
-      <ReloadArrows />
-      <GridGraph Node={Node} nodes={nodes} nodeClassName={classNames(styles.bubbleContainer, nodeClassName)} {...rest}>
-        {children}
-      </GridGraph>
-    </Xwrapper>
+    <GridGraph Node={Node} nodes={nodes} nodeClassName={classNames(styles.bubbleContainer, nodeClassName)} {...rest}>
+      {children}
+    </GridGraph>
   );
-}
-
-// this will be fixed in react-xarrows v3.0.0
-function ReloadArrows() {
-  const updateXarrow = useXarrow();
-
-  useEffect(() => {
-    updateXarrow();
-  }, []);
-
-  return null;
 }
