+++ community/ui/graph/edge/index.ts
@@ -1,2 +1,5 @@
 export { Edge } from './edge';
 export type { EdgeProps } from './edge';
+
+export { ArrowAutoReloader } from './auto-reloader';
+export type { ArrowAutoReloaderProps } from './auto-reloader';
